move param.c into gadget.
